refactor: 패키지 구조 변경
1. guarantee, institute 패키지를 model 패키지의 하위 패키지로 이동시킨다
