+++ src/main/java/com/ehem/kakaopay/model/guarantee/domain/vo/Amount.java
@@ -1,4 +1,4 @@
-package com.ehem.kakaopay.guarantee.domain.vo;
+package com.ehem.kakaopay.model.guarantee.domain.vo;
 
 import lombok.AccessLevel;
 import lombok.EqualsAndHashCode;
