+++ src/main/java/com/ehem/kakaopay/model/institute/domain/vo/InstituteType.java
@@ -1,10 +1,8 @@
-package com.ehem.kakaopay.institute.domain.vo;
+package com.ehem.kakaopay.model.institute.domain.vo;
 
 import lombok.Getter;
 
 import javax.persistence.Convert;
-import java.util.Arrays;
-import java.util.Collections;
 import java.util.stream.Stream;
 
 @Getter
