--- src/test/java/com/ehem/kakaopay/institute/domain/InstituteTest.java
@@ -1,7 +1,0 @@
-package com.ehem.kakaopay.institute.domain;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class InstituteTest {
-
-}
