+++ src/test/java/com/ehem/kakaopay/model/guarantee/domain/GuaranteeTest.java
@@ -1,10 +1,10 @@
-package com.ehem.kakaopay.guarantee.domain;
+package com.ehem.kakaopay.model.guarantee.domain;
 
-import com.ehem.kakaopay.guarantee.domain.vo.Amount;
-import com.ehem.kakaopay.institute.domain.Institute;
-import com.ehem.kakaopay.institute.domain.vo.InstituteType;
-import com.ehem.kakaopay.guarantee.domain.vo.Month;
-import com.ehem.kakaopay.guarantee.domain.vo.Year;
+import com.ehem.kakaopay.model.guarantee.domain.vo.Amount;
+import com.ehem.kakaopay.model.institute.domain.Institute;
+import com.ehem.kakaopay.model.institute.domain.vo.InstituteType;
+import com.ehem.kakaopay.model.guarantee.domain.vo.Month;
+import com.ehem.kakaopay.model.guarantee.domain.vo.Year;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
