+++ src/test/java/com/ehem/kakaopay/model/institute/domain/InstituteTest.java
@@ -1,0 +1,5 @@
+package com.ehem.kakaopay.model.institute.domain;
+
+class InstituteTest {
+
+}
